接RxJava和Retrofit2准备

+++ app/src/main/java/android/kevin/cn/ks/activity/StatisticsActivity.java
@@ -106,7 +106,7 @@ public class StatisticsActivity extends BaseActivity {
 
     void fillList() {
         while(list.size() <= 3) {
-            list.add(new PlanStatistics(1L, new Plan(Integer.parseInt(list.size() + ""), "测试", "描述"), 5));
+            list.add(new PlanStatistics(1L, new Plan(Long.parseLong(list.size() + ""), "测试", "描述"), 5));
         }
     }
